+++ src/Menu.js
@@ -4,34 +4,34 @@ import Tap from 'tap.js';
 
 export default class Menu extends React.Component {
   componentWillMount() {
-    this.props.manager.menu = this;
+    this.context.ambManager.menu = this;
 
     this.isListeningForTap = false;
     this.tapHandler = (e) => {
       if (ReactDOM.findDOMNode(this).contains(e.target)) return;
-      if (ReactDOM.findDOMNode(this.props.manager.button).contains(e.target)) return;
-      this.props.manager.closeMenu();
+      if (ReactDOM.findDOMNode(this.context.ambManager.button).contains(e.target)) return;
+      this.context.ambManager.closeMenu();
     }
   }
 
   componentWillUpdate() {
-    const { manager } = this.props;
-    if (manager.isOpen && !this.isListeningForTap) {
+    const { ambManager } = this.context;
+    if (ambManager.isOpen && !this.isListeningForTap) {
       this.addTapListeners();
-    } else if (!manager.isOpen && this.isListeningForTap) {
+    } else if (!ambManager.isOpen && this.isListeningForTap) {
       this.removeTapListeners();
     }
 
-    if (!manager.isOpen) {
-      // Clear the manager's items, so they
+    if (!ambManager.isOpen) {
+      // Clear the ambManager's items, so they
       // can be reloaded next time this menu opens
-      manager.menuItems = [];
+      ambManager.menuItems = [];
     }
   }
 
   componentWillUnmount() {
     this.removeTapListeners();
-    this.props.manager.powerDown();
+    this.context.ambManager.destroy();
   }
 
   addTapListeners() {
@@ -50,13 +50,14 @@ export default class Menu extends React.Component {
   }
 
   render() {
-    const { manager, children, tag, className, id } = this.props;
+    const { children, tag, className, id, style } = this.props;
+    const { ambManager } = this.context;
 
     const childrenToRender = (() => {
       if (typeof children === 'function') {
-        return children({ isOpen: manager.isOpen });
+        return children({ isOpen: ambManager.isOpen });
       }
-      if (manager.isOpen) return children;
+      if (ambManager.isOpen) return children;
       return false;
     })();
 
@@ -65,21 +66,26 @@ export default class Menu extends React.Component {
     return React.createElement(tag, {
       className,
       id,
-      onKeyDown: manager.handleMenuKey,
+      style,
+      onKeyDown: ambManager.handleMenuKey,
       role: 'menu',
-      onBlur: manager.handleBlur,
+      onBlur: ambManager.handleBlur,
     }, childrenToRender);
   }
 }
 
 Menu.propTypes = {
-  children: PropTypes.oneOfType([PropTypes.func, PropTypes.element]).isRequired,
-  manager: PropTypes.object.isRequired,
+  children: PropTypes.oneOfType([PropTypes.func, PropTypes.node]).isRequired,
   id: PropTypes.string,
   className: PropTypes.string,
+  style: PropTypes.object,
   tag: PropTypes.string,
 };
 
 Menu.defaultProps = {
   tag: 'div',
+};
+
+Menu.contextTypes = {
+  ambManager: PropTypes.object.isRequired,
 };