+++ src/index.js
@@ -1,0 +1,11 @@
+import Wrapper from './Wrapper';
+import Button from './Button';
+import Menu from './Menu';
+import MenuItem from './MenuItem';
+
+export default {
+  Wrapper,
+  Button,
+  Menu,
+  MenuItem,
+};